move sine and lcs_coords_from_ts functions to welding module
