+++ weldx/util.py
@@ -21,7 +21,6 @@ from scipy.spatial.transform import Slerp
 
 from weldx.constants import WELDX_QUANTITY as Q_
 from weldx.constants import WELDX_UNIT_REGISTRY as ureg
-from weldx.core import MathematicalExpression, TimeSeries
 
 if TYPE_CHECKING:  # pragma: no cover
     import weldx.transformations as tf
@@ -226,74 +225,6 @@ def inherit_docstrings(cls):
                         f"could not derive docstring for {cls}.{name}", stacklevel=1
                     )
     return cls
-
-
-def sine(
-    f: Union[pint.Quantity, str],
-    amp: Union[pint.Quantity, str],
-    bias: Union[pint.Quantity, str] = None,
-    phase: Union[pint.Quantity, str] = Q_(0, "rad"),
-) -> TimeSeries:
-    """Create a simple sine TimeSeries from quantity parameters.
-
-    f(t) = amp*sin(f*t+phase)+bias
-
-    Parameters
-    ----------
-    f :
-        Frequency of the sine (in Hz)
-    amp :
-        Sine amplitude
-    bias :
-        function bias
-    phase :
-        phase shift
-
-    Returns
-    -------
-    TimeSeries
-
-    """
-    if bias is None:
-        amp = Q_(amp)
-        bias = 0.0 * amp.u
-    expr_string = "a*sin(o*t+p)+b"
-    parameters = {"a": amp, "b": bias, "o": Q_(2 * np.pi, "rad") * Q_(f), "p": phase}
-    expr = MathematicalExpression(expression=expr_string, parameters=parameters)
-    return TimeSeries(expr)
-
-
-@deprecated(
-    "0.4.1",
-    "0.5.0",
-    "The 'LocalCoordinateSystem' now supports 'TimeSeries' as coordinates rendering "
-    "this function obsolete.",
-)
-def lcs_coords_from_ts(
-    ts: TimeSeries, time: Union[pd.DatetimeIndex, pint.Quantity]
-) -> xr.DataArray:
-    """Create translation coordinates from a TimeSeries at specific timesteps.
-
-    Parameters
-    ----------
-    ts:
-        TimeSeries that describes the coordinate motion as a 3D vector.
-    time
-        Timestamps used for interpolation.
-        TODO: add support for pd.DateTimeindex as well
-
-    Returns
-    -------
-    xarray.DataArray :
-        A DataArray with correctly labeled dimensions to be used for LCS creation.
-
-    """
-    ts_data = ts.interp_time(time=time).data_array
-    # assign vector coordinates and convert to mm
-    ts_data = ts_data.rename({"dim_1": "c"}).assign_coords({"c": ["x", "y", "z"]})
-    ts_data.data = ts_data.data.to("mm").magnitude
-    ts_data["time"] = pd.TimedeltaIndex(ts_data["time"].data)
-    return ts_data
 
 
 def is_column_in_matrix(column, matrix) -> bool:
